feat: 로그인 후 커뮤니티 각 기능별 인가와 추가 기능 구현완료

+++ src/pages/Community/components/userInfoComponents/CommentHistory.tsx
@@ -35,13 +35,25 @@ interface PostDetail {
   user: UserDetail;
 }
 
-export default function CommentHistory() {
+type CommentHistoryProps = {
+  profileId: number | null | undefined;
+};
+
+interface Headers {
+  'Content-Type': string;
+  Authorization?: string;
+  [key: string]: string | undefined;
+}
+
+export default function CommentHistory({ profileId }: CommentHistoryProps) {
   const [comments, setComments] = useState<CommentDetail[]>();
   const [checked, setChecked] = useState<number[]>([]);
   const [commentCount, setCommentCount] = useState<number>(0);
   const [page, setPage] = useState<number>(1);
 
   const navigate = useNavigate();
+  const loginUserToken = localStorage.getItem('accessToken');
+  const loginUserId = Number(localStorage.getItem('id'));
 
   const dateParsing = (date: string): [string, boolean] => {
     const theDate = new Date(date);
@@ -60,21 +72,34 @@ export default function CommentHistory() {
   };
 
   useEffect(() => {
-    fetch(`http://pien.kr:4000/community/reply/user/1?page=${page}&number=15`, {
-      headers: {
-        'Content-Type': 'application/json;charset=utf-8',
-      },
+    const headers: Headers = {
+      'Content-Type': 'application/json;charset=utf-8',
+    };
+
+    if (loginUserToken) {
+      headers.Authorization = `Bearer ${loginUserToken}`;
+    } else {
+      delete headers.Authorization;
+    }
+
+    fetch(`http://pien.kr:4000/community/reply/user/${profileId}?page=${page}&number=15`, {
+      headers: Object.entries(headers).map(([key, value]) => [key, value || '']),
     })
       .then((res) => res.json())
       .then((data) => {
-        setCommentCount(data.number);
-        setComments(
-          data.replies.map((el: CommentDetail) => ({
-            ...el,
-            created_at: dateParsing(el.created_at)[0],
-            isItNew: dateParsing(el.created_at)[1],
-          })),
-        );
+        if (data.isSuccess) {
+          setCommentCount(data.data.number);
+          setComments(
+            data.data.replies.map((el: CommentDetail) => ({
+              ...el,
+              created_at: dateParsing(el.created_at)[0],
+              isItNew: dateParsing(el.created_at)[1],
+            })),
+          );
+        } else {
+          alert(data.message);
+          navigate('/community/list');
+        }
       });
   }, [page]);
 
@@ -100,38 +125,53 @@ export default function CommentHistory() {
   };
 
   const deleteComment = () => {
-    if (window.confirm('댓글을 삭제하시겠습니까?') === true) {
-      fetch(`http://pien.kr:4000/community/reply/`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json;charset=utf-8',
-          Authorization:
-            'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJraXN1azYyM0BuYXZlci5jb20iLCJpYXQiOjE2NzM5Mzg4OTUsImV4cCI6MTY3Mzk0MDY5NX0.VWzQ1BIRwbrdAn1RLcmHol8lTtZf4Yx5we2pLpzQr3U',
-        },
-        body: JSON.stringify({ replyId: checked }),
-      })
-        .then((res) => res.json())
-        .then((data) => {
-          if (data.status === true) {
-            fetch(`http://pien.kr:4000/community/reply/user/1?page=${page}&number=15`, {
-              headers: {
-                'Content-Type': 'application/json;charset=utf-8',
-              },
-            })
-              .then((res) => res.json())
-              .then((data) => {
-                setCommentCount(data.number);
-                setComments(
-                  data.replies.map((el: CommentDetail) => ({
-                    ...el,
-                    created_at: dateParsing(el.created_at)[0],
-                    isItNew: dateParsing(el.created_at)[1],
-                  })),
-                );
-                setChecked([]);
-              });
-          }
-        });
+    if (!loginUserToken) {
+      if (window.confirm('로그인 후 이용가능합니다. 로그인 하시겠습니까?') === true) {
+        navigate('login');
+      }
+    } else {
+      const headers: Headers = {
+        'Content-Type': 'application/json;charset=utf-8',
+      };
+
+      if (loginUserToken) {
+        headers.Authorization = `Bearer ${loginUserToken}`;
+      } else {
+        delete headers.Authorization;
+      }
+
+      if (window.confirm('댓글을 삭제하시겠습니까?') === true) {
+        fetch(`http://pien.kr:4000/community/reply/`, {
+          method: 'DELETE',
+          headers: Object.entries(headers).map(([key, value]) => [key, value || '']),
+          body: JSON.stringify({ replyId: checked }),
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.isSuccess) {
+              fetch(`http://pien.kr:4000/community/reply/user/1?page=${page}&number=15`, {
+                headers: Object.entries(headers).map(([key, value]) => [key, value || '']),
+              })
+                .then((res) => res.json())
+                .then((data) => {
+                  if (data.isSuccess) {
+                    setCommentCount(data.data.number);
+                    setComments(
+                      data.data.replies.map((el: CommentDetail) => ({
+                        ...el,
+                        created_at: dateParsing(el.created_at)[0],
+                        isItNew: dateParsing(el.created_at)[1],
+                      })),
+                    );
+                    setChecked([]);
+                  } else {
+                    alert(data.message);
+                    navigate('/community/list');
+                  }
+                });
+            }
+          });
+      }
     }
     return null;
   };
@@ -149,15 +189,17 @@ export default function CommentHistory() {
         {comments?.map((el) => (
           <CommentBox key={el.id}>
             <CommentInner>
-              <CheckBox>
-                <input
-                  type='checkBox'
-                  id={String(el.id)}
-                  checked={checked.includes(el.id)}
-                  onChange={(event) => checkedChange(event)}
-                  readOnly
-                />
-              </CheckBox>
+              {profileId === loginUserId ? (
+                <CheckBox>
+                  <input
+                    type='checkBox'
+                    id={String(el.id)}
+                    checked={checked.includes(el.id)}
+                    onChange={(event) => checkedChange(event)}
+                    readOnly
+                  />
+                </CheckBox>
+              ) : null}
               <Comment
                 onClick={() => (el.post.isPublished ? navigate(`/community/${el.post.id}`) : null)}
                 isPublished={el.post.isPublished}
@@ -178,15 +220,29 @@ export default function CommentHistory() {
       </List>
       <ButtonBox>
         <SelectAll>
-          <CheckAll onClick={checkAll}>
-            <input type='checkBox' checked={checked.length === comments?.length} readOnly />
-            <div>전체선택</div>
-          </CheckAll>
+          {profileId === loginUserId && commentCount !== 0 ? (
+            <CheckAll onClick={checkAll}>
+              <input type='checkBox' checked={checked.length === comments?.length} readOnly />
+              <div>전체선택</div>
+            </CheckAll>
+          ) : null}
         </SelectAll>
         <Pages page={page} setPage={setPage} postNumber={commentCount} limit={15} />
         <DeleteAndWrite>
-          <DeleteBtn onClick={deleteComment}>삭제</DeleteBtn>
-          <WriteBtn onClick={() => navigate('/community/posting')}>글쓰기</WriteBtn>
+          {profileId === loginUserId ? <DeleteBtn onClick={deleteComment}>삭제</DeleteBtn> : null}
+          <WriteBtn
+            onClick={() => {
+              if (!loginUserToken) {
+                if (window.confirm('로그인 후 이용가능합니다. 로그인 하시겠습니까?') === true) {
+                  navigate('/login');
+                }
+              } else {
+                navigate('/community/posting');
+              }
+            }}
+          >
+            글쓰기
+          </WriteBtn>
         </DeleteAndWrite>
       </ButtonBox>
     </OuterBox>
